Try Memory Display

+++ Source/Controller/controller.hpp
@@ -59,7 +59,7 @@ private:
     QMap<QString, QByteArray> incomingDataToDownload;
     int currentNodeLoaded = 0;
     int connectedNodes = 0;
-    int spaceUsed = 0;
+    quint64 spaceUsed = 0;
     quint64 totalMemory; 
 };
